Add internationalization and setup page with language support
- Implemented localization using Flutter's internationalization framework
- Created language provider for managing app language settings
- Added language selector widget with English, Chinese, and Polish translations
- Developed setup page with language, notifications, security, and backup sections
- Updated main app configuration to support multiple languages
- Added language ARB files for different locales
- Integrated language persistence using secure storage

+++ lib/home.dart
@@ -296,8 +296,8 @@ class _HomeScreenState extends State<HomeScreen> {
         final temperature = sensor['temperature']?.toString() ?? '--';
         final battery = sensor['battery']?.toString() ?? '--';
         final timestamp = sensor['timestamp'] ?? '--';
-        final minTemp = sensor['min_temp'] ?? 0;
-        final maxTemp = sensor['max_temp'] ?? 5;
+        final minTemp = (sensor['min_temp'] ?? 0).toDouble();
+        final maxTemp = (sensor['max_temp'] ?? 5).toDouble();
 
         return Container(
           width: 150,
@@ -492,7 +492,7 @@ class _HomeScreenState extends State<HomeScreen> {
 
       return LineChartBarData(
         spots: data.asMap().entries.map((e) {
-          return FlSpot(e.key.toDouble(), e.value);
+          return FlSpot(e.key.toDouble(), e.value.toDouble());
         }).toList(),
         isCurved: true,
         color: color,
